requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.2.0
anyio==4.2.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.11.0
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.6.2
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cryptography==41.0.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
executing==0.8.3
fastjsonschema==2.16.2
flatbuffers==2.0
fonttools==4.51.0
frozenlist==1.4.0
gast==0.4.0
google-auth==2.6.0
google-auth-oauthlib==0.5.2
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.9.0
idna==3.7
ipykernel==6.28.0
ipython==8.20.0
jedi==0.18.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.6
jsonpickle==2.2.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
nltk==3.8.1
notebook==7.0.8
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pillow==10.3.0
pip==23.3.1
platformdirs==3.10.0
pooch==1.7.0
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==2.0.0
Pygments==2.15.1
PyJWT==2.4.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.7.2
scikit-learn==1.1.3
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.2
setuptools==68.2.2
simplegeneric==0.8.1
six==1.16.0
sklearn-pandas==2.2.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
termcolor==2.1.0
terminado==0.17.1
threadpoolctl==2.2.0
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.2
traitlets==5.7.1
typing_extensions==4.9.0
tzdata==2023.3
unicodedata2==15.1.0
urllib3==2.1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.3.8
wheel==0.35.1
wordcloud==1.9.3
wrapt==1.14.1
yarl==1.9.3